+++ lib/ahoy/stores/base_store.rb
@@ -64,6 +64,20 @@ module Ahoy
         ahoy.visit_properties
       end
 
+      def set_visit_properties(visit)
+        keys = visit_properties.keys
+        keys -= Ahoy::VisitProperties::LOCATION_KEYS if Ahoy.geocode != true
+        keys.each do |key|
+          visit.send(:"#{key}=", visit_properties[key]) if visit.respond_to?(:"#{key}=") && visit_properties[key]
+        end
+      end
+
+      def geocode(visit)
+        if Ahoy.geocode == :async
+          Ahoy::GeocodeJob.perform_later(visit)
+        end
+      end
+
     end
   end
 end